save to complete at home

+++ include/Utils.hpp
@@ -135,7 +135,7 @@ namespace Utils {
 	{
 		int occur = 0;
 
-		for (int i = 0; i < str.length(); i++) {
+		for (size_t i = 0; i < str.length(); i++) {
 			if (str[i] == ch) {
 				occur += 1;
 			}
